requirements.txt
flask
pydantic
openai
google-cloud-vision
python-dotenv
rich
gunicorn